+++ demo/js/fancyDemo.jsx
@@ -9,7 +9,7 @@ class FancyDemo extends React.Component {
         <TabList>
           <ul className='FancyTabs-tablist'>
             <li className='FancyTabs-tablistItem'>
-              <Tab tabId='1' className='FancyTabs-tab'>
+              <Tab tabId='t1' className='FancyTabs-tab'>
                 {demoTab.bind(null, (
                   <div>
                     <span className='FancyTabs-tabIcon FancyTabs-tabIcon--map' />
@@ -21,7 +21,7 @@ class FancyDemo extends React.Component {
               </Tab>
             </li>
             <li className='FancyTabs-tablistItem'>
-              <Tab tabId='2' className='FancyTabs-tab'>
+              <Tab tabId='t2' className='FancyTabs-tab'>
                 {demoTab.bind(null, (
                   <div>
                     <span className='FancyTabs-tabIcon FancyTabs-tabIcon--megaphone' />
@@ -33,7 +33,7 @@ class FancyDemo extends React.Component {
               </Tab>
             </li>
             <li className='FancyTabs-tablistItem'>
-              <Tab tabId='3' className='FancyTabs-tab'>
+              <Tab tabId='t3' className='FancyTabs-tab'>
                 {demoTab.bind(null, (
                   <div>
                     <span className='FancyTabs-tabIcon FancyTabs-tabIcon--trophy' />
@@ -47,17 +47,17 @@ class FancyDemo extends React.Component {
           </ul>
         </TabList>
         <div className='FancyTabs-panel'>
-          <TabPanel tabId='1'>
+          <TabPanel tabId='t1'>
             <div className='FancyTabs-panelInner'>
               Lorem <a href='#'>ipsum</a> dolor sit amet, consectetur adipisicing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.
             </div>
           </TabPanel>
-          <TabPanel tabId='2'>
+          <TabPanel tabId='t2'>
             <div className='FancyTabs-panelInner'>
               Ut <a href='#'>enim</a> ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.
             </div>
           </TabPanel>
-          <TabPanel tabId='3'>
+          <TabPanel tabId='t3'>
             <div className='FancyTabs-panelInner'>
               Duis <a href='#'>aute</a> irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.
             </div>