+++ demo/js/statelessDemo.jsx
@@ -5,7 +5,7 @@ import { Wrapper, Tab, TabList, TabPanel} from '../../src';
 const tabData = [
   {
     title: 'one',
-    id: '1',
+    id: 't1',
     content: (
       <div>
         Lorem <a href='#'>ipsum</a> dolor sit amet, consectetur adipisicing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.
@@ -14,7 +14,7 @@ const tabData = [
   },
   {
     title: 'two',
-    id: '2',
+    id: 't2',
     content: (
       <div>
         Ut <a href='#'>enim</a> ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.
@@ -23,7 +23,7 @@ const tabData = [
   },
   {
     title: 'three',
-    id: '3',
+    id: 't3',
     content: (
       <div>
         Duis <a href='#'>aute</a> irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.