+++ src/TabPanel.js
@@ -33,7 +33,7 @@ export default class TabPanel extends React.Component {
 
     return React.createElement(props.tag, {
       className: props.className,
-      id: props.tabId,
+      id: props.id,
       onKeyDown: this.handleKeyDown.bind(this),
       role: 'tabpanel',
       style: props.style,
@@ -49,6 +49,7 @@ TabPanel.propTypes = {
   ]).isRequired,
   tabId: PropTypes.string.isRequired,
   className: PropTypes.string,
+  id: PropTypes.string,
   style: PropTypes.object,
   tag: PropTypes.string,
 };